atb-tracker/backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
PyJWT==2.8.0 